Addressed several bugs and added support for reasoning when graphing.
Reasoning is used to determine super-classes and individual types beyond
what is specified. It was primarily added to address when multiple domains
and/or ranges are defined for a property - in which case, a referenced
individual is inferred to be of ALL the specified types (an intersection
of all the specified classes). The models, GUI and test were updated to
allow reasoning to be specified in a graphing request.

The following bugs were addressed:
* #7 (allow for use of less than/greater than signs in labels)
* #8 (correctly show how multiple domains and ranges for a property are
  interpreted with reasoning)
* #29 (collapse/expand customization options in order to reduce
  screen real-estate)
* Fixed test data where ", " was used in a label (but it should have
  been the text, COMMA_SPACE
* Updates required after supporting multiple domain and range classes:
  * UML graph of individuals added a "Restriction:" note, each time that there
    was more than 1 type
  * VOWL did not correctly show the relation to a specificThing for the second
    of a multi-range property (it referenced the specific Thing for the first
    range property)
  * VOWL graph could have an owl:Thing node with no refs, if all the references
    were updated due to class splitting. This node is now removed.
* In GUI testing, found that "collapse edges" was displayed for a UML Individual
  graph, which was not valid. Also found that "collapse edges" was not displayed
  for a Graffoo  "both" class & property graph. Modified the javascript to check
  for (UML & not Individual) OR (Graffoo & (property or both)).

Also, corrected various comments and java docs.

Removed ./ontograph-1.0.1 directory which was an artifact of testing.